+++ src/pages/axios-test/AxiosTest.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import AxiosExample from '@/components/axios-test/AxiosExample';
+
+const AxiosTest: React.FC = () => {
+  return <AxiosExample />;
+};
+
+export default AxiosTest;
